[AIRFLOW-6721] Move Hive MetastorePartitionSensor test
Move the MetastorePartitionSensor test from the operators module
into a new module and use the shared test environment.
